Refactor the pipe modules to fit semantics
